--- core/src/main/java/com/es/core/cart/Cart.java
@@ -1,42 +1,0 @@
-package com.es.core.cart;
-
-import java.math.BigDecimal;
-import java.util.ArrayList;
-import java.util.List;
-
-public class Cart {
-
-    private List<CartItem> cartItems;
-    private Long totalQuantity;
-    private BigDecimal totalCost;
-
-    public Cart() {
-        cartItems = new ArrayList<>();
-        totalQuantity = 0L;
-        totalCost = BigDecimal.ZERO;
-    }
-
-    public List<CartItem> getCartItems() {
-        return cartItems;
-    }
-
-    public void setCartItems(List<CartItem> cartItems) {
-        this.cartItems = cartItems;
-    }
-
-    public Long getTotalQuantity() {
-        return totalQuantity;
-    }
-
-    public void setTotalQuantity(Long totalQuantity) {
-        this.totalQuantity = totalQuantity;
-    }
-
-    public BigDecimal getTotalCost() {
-        return totalCost;
-    }
-
-    public void setTotalCost(BigDecimal totalCost) {
-        this.totalCost = totalCost;
-    }
-}
