--- core/src/main/java/com/es/core/filter/FilterService.java
@@ -1,6 +1,0 @@
-package com.es.core.filter;
-
-public interface FilterService {
-    String checkFieldValue(String field);
-    String checkOrderValue(String order);
-}
