--- core/src/main/java/com/es/core/filter/FilterServiceImpl.java
@@ -1,37 +1,0 @@
-package com.es.core.filter;
-
-import com.es.core.sortenum.SortField;
-import com.es.core.sortenum.SortOrder;
-import org.springframework.stereotype.Service;
-
-@Service
-public class FilterServiceImpl implements FilterService {
-
-    @Override
-    public String checkFieldValue(String field) {
-        if (field != null) {
-            try {
-                SortField sortField = SortField.valueOf(field);
-                return sortField.name();
-            } catch (IllegalArgumentException e) {
-                return null;
-            }
-        } else {
-            return null;
-        }
-    }
-
-    @Override
-    public String checkOrderValue(String order) {
-        if (order != null) {
-            try {
-                SortOrder sortOrder = SortOrder.valueOf(order);
-                return sortOrder.name();
-            } catch (IllegalArgumentException e) {
-                return null;
-            }
-        } else {
-            return null;
-        }
-    }
-}
