+++ core/src/main/java/com/es/core/model/color/Color.java
@@ -1,0 +1,22 @@
+package com.es.core.model.color;
+
+public class Color {
+    private Long id;
+    private String code;
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getCode() {
+        return code;
+    }
+
+    public void setCode(final String code) {
+        this.code = code;
+    }
+}
