+++ core/src/main/java/com/es/core/model/color/ColorDao.java
@@ -1,0 +1,9 @@
+package com.es.core.model.color;
+
+import java.util.Optional;
+
+public interface ColorDao {
+    Optional<Color> get(final Long key);
+
+    void save(final Color color);
+}
