+++ core/src/main/java/com/es/core/model/order/OrderItem.java
@@ -1,6 +1,6 @@
 package com.es.core.model.order;
 
-import com.es.core.cart.CartItem;
+import com.es.core.model.cart.CartItem;
 import com.es.core.model.phone.Phone;
 
 public class OrderItem {
