--- core/src/main/java/com/es/core/model/phone/ColorDao.java
@@ -1,8 +1,0 @@
-package com.es.core.model.phone;
-
-import java.util.Optional;
-
-public interface ColorDao {
-    Optional<Color> get(final Long key);
-    void save(final Color color);
-}
