+++ core/src/main/java/com/es/core/model/phone/Phone.java
@@ -1,5 +1,7 @@
 package com.es.core.model.phone;
 
+import com.es.core.model.color.Color;
+
 import java.math.BigDecimal;
 import java.util.Collections;
 import java.util.Date;
