+++ core/src/main/java/com/es/core/model/stock/Stock.java
@@ -1,0 +1,33 @@
+package com.es.core.model.stock;
+
+import com.es.core.model.phone.Phone;
+
+public class Stock {
+    private Phone phone;
+    private Integer stock;
+    private Integer reserved;
+
+    public Phone getPhone() {
+        return phone;
+    }
+
+    public void setPhone(Phone phone) {
+        this.phone = phone;
+    }
+
+    public Integer getStock() {
+        return stock;
+    }
+
+    public void setStock(Integer stock) {
+        this.stock = stock;
+    }
+
+    public Integer getReserved() {
+        return reserved;
+    }
+
+    public void setReserved(Integer reserved) {
+        this.reserved = reserved;
+    }
+}
