+++ core/src/main/java/com/es/core/model/stock/StockDao.java
@@ -1,0 +1,9 @@
+package com.es.core.model.stock;
+
+import java.util.Optional;
+
+public interface StockDao {
+    Optional<Stock> get(final Long key);
+
+    void update(final Long key, final Long stock, final Long reserved);
+}
