+++ core/src/main/java/com/es/core/service/filter/FilterService.java
@@ -1,0 +1,7 @@
+package com.es.core.service.filter;
+
+public interface FilterService {
+    String checkFieldValue(String field);
+
+    String checkOrderValue(String order);
+}
