+++ core/src/main/resources/context/applicationContext-core.xml
@@ -16,6 +16,10 @@
         <property name="password" value="${db.password}" />
     </bean>
 
+    <bean id="userDetailsService" class="org.springframework.security.core.userdetails.jdbc.JdbcDaoImpl">
+        <property name="dataSource" ref="dataSource"/>
+    </bean>
+
     <bean id="jdbcTemplate" class="org.springframework.jdbc.core.JdbcTemplate">
         <property name="dataSource" ref="dataSource" />
     </bean>
