+++ core/src/test/java/OrderServiceImplTest.java
@@ -1,14 +1,13 @@
-import com.es.core.cart.Cart;
-import com.es.core.cart.CartItem;
+import com.es.core.model.cart.Cart;
+import com.es.core.model.cart.CartItem;
 import com.es.core.exception.NoElementWithSuchIdException;
-import com.es.core.model.order.Order;
 import com.es.core.model.order.OrderDao;
 import com.es.core.model.order.OrderDataDTO;
 import com.es.core.model.phone.Phone;
-import com.es.core.model.phone.Stock;
-import com.es.core.model.phone.StockDao;
-import com.es.core.order.OrderService;
-import com.es.core.order.OrderServiceImpl;
+import com.es.core.model.stock.Stock;
+import com.es.core.model.stock.StockDao;
+import com.es.core.service.order.OrderService;
+import com.es.core.service.order.OrderServiceImpl;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
