+++ web/src/main/java/com/es/phoneshop/web/controller/AjaxCartController.java
@@ -1,10 +1,10 @@
 package com.es.phoneshop.web.controller;
 
-import com.es.core.cart.Cart;
-import com.es.core.cart.CartService;
 import com.es.core.exception.EmptyDatabaseArgumentException;
 import com.es.core.exception.OutOfStockException;
+import com.es.core.model.cart.Cart;
 import com.es.core.model.phone.PhoneDTO;
+import com.es.core.service.cart.CartService;
 import com.es.core.validator.HandlingInfo;
 import com.es.core.validator.ResponseHandlingInfo;
 import com.es.core.validator.SuccessfulHandlingInfo;
@@ -54,6 +54,5 @@ public class AjaxCartController {
         } catch (Exception e) {
             return ResponseEntity.status(500).build();
         }
-
     }
 }
