+++ web/src/main/java/com/es/phoneshop/web/controller/pages/ProductDetailsPageController.java
@@ -1,12 +1,11 @@
 package com.es.phoneshop.web.controller.pages;
 
-import com.es.core.cart.CartService;
 import com.es.core.exception.NoElementWithSuchIdException;
 import com.es.core.model.phone.Phone;
-import com.es.core.model.phone.PhoneDao;
+import com.es.core.service.cart.CartService;
+import com.es.core.service.phone.PhoneService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.env.Environment;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.ExceptionHandler;
@@ -16,7 +15,6 @@ import org.springframework.web.bind.annotation.RequestMethod;
 
 import javax.annotation.Resource;
 import javax.servlet.http.HttpSession;
-import java.util.Optional;
 
 @Controller
 @RequestMapping(value = "/productDetails/{id}")
@@ -26,7 +24,7 @@ public class ProductDetailsPageController {
     private Environment env;
 
     @Resource
-    private PhoneDao jdbcPhoneDao;
+    private PhoneService phoneService;
 
     @Resource
     private CartService cartService;
@@ -36,14 +34,9 @@ public class ProductDetailsPageController {
 
     @RequestMapping(method = RequestMethod.GET)
     public String showProductDetails(@PathVariable("id") String id, Model model) throws NoElementWithSuchIdException {
-        Optional<Phone> currentPhone;
-        try {
-            currentPhone = jdbcPhoneDao.get(Long.valueOf(id));
-            model.addAttribute("cart", cartService.getCart(httpSession));
-            currentPhone.ifPresent(phone -> model.addAttribute("phone", phone));
-        } catch (NumberFormatException | EmptyResultDataAccessException e) {
-            throw new NoElementWithSuchIdException(id);
-        }
+        Phone currentPhone = phoneService.getPhone(id);
+        model.addAttribute("cart", cartService.getCart(httpSession));
+        model.addAttribute("phone", currentPhone);
         return "productDetails";
     }
