+++ web/src/main/java/com/es/phoneshop/web/controller/pages/admin/OrdersDetailsPageController.java
@@ -3,12 +3,11 @@ package com.es.phoneshop.web.controller.pages.admin;
 import com.es.core.exception.IncorrectUrlParameterException;
 import com.es.core.exception.NoElementWithSuchIdException;
 import com.es.core.model.order.Order;
-import com.es.core.model.order.OrderDao;
 import com.es.core.model.order.OrderStatus;
+import com.es.core.service.order.OrderService;
 import org.apache.commons.lang3.EnumUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.env.Environment;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -19,38 +18,27 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 
 
 import javax.annotation.Resource;
-import java.util.Optional;
 
 @Controller
 @RequestMapping(value = "/admin/orders/{id}")
 public class OrdersDetailsPageController {
 
     @Resource
-    private OrderDao jdbcOrderDao;
+    private OrderService orderService;
 
     @Autowired
     private Environment env;
 
     @RequestMapping(method = RequestMethod.GET)
     public String getOrder(@PathVariable("id") String id, Model model) {
-        Optional<Order> currentOrder;
-        try {
-            currentOrder = jdbcOrderDao.get(Long.valueOf(id));
-            currentOrder.ifPresent(order -> model.addAttribute("order", order));
-        } catch (NumberFormatException | EmptyResultDataAccessException e ) {
-            throw new NoElementWithSuchIdException(id);
-        }
+        Order order = orderService.getOrder(id);
+        model.addAttribute("order", order);
         return "adminOrderDetails";
     }
 
     @RequestMapping(method = RequestMethod.POST)
     public String changeStatus(@RequestParam(required = false) String status, @PathVariable("id") Long id) {
-        Optional<Order> currentOrder = jdbcOrderDao.get(id);
-        if (currentOrder.isPresent()) {
-            updateStatus(status, currentOrder.get());
-        } else {
-            throw new NoElementWithSuchIdException(id.toString());
-        }
+        updateStatus(status, orderService.getOrder(id.toString()));
         return "redirect:/admin/orders/" + id;
     }
 
@@ -58,7 +46,7 @@ public class OrdersDetailsPageController {
         status = status.toUpperCase();
         if (EnumUtils.isValidEnum(OrderStatus.class, status)) {
             order.setStatus(OrderStatus.valueOf(status));
-            jdbcOrderDao.updateStatus(order.getId(), order.getStatus());
+            orderService.updateStatus(order.getId(), order.getStatus());
         } else {
             throw new IncorrectUrlParameterException(status);
         }
