+++ web/src/main/java/com/es/phoneshop/web/controller/pages/admin/OrdersPageController.java
@@ -1,6 +1,6 @@
 package com.es.phoneshop.web.controller.pages.admin;
 
-import com.es.core.model.order.OrderDao;
+import com.es.core.service.order.OrderService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.env.Environment;
 import org.springframework.stereotype.Controller;
@@ -19,7 +19,7 @@ public class OrdersPageController {
     private Environment env;
 
     @Resource
-    private OrderDao orderDao;
+    private OrderService orderService;
 
     @RequestMapping(method = RequestMethod.GET)
     public String showOrders(@RequestParam(required = false) Long page, Model model) {
@@ -27,11 +27,12 @@ public class OrdersPageController {
             page = 1L;
         }
         Long quantityOnPage = Long.parseLong(env.getProperty("value.quantityOnPage"));
-        Long orderQuantity = orderDao.count();
+        Long orderQuantity = orderService.countAllOrders();
         Long pagesQuantity = orderQuantity / quantityOnPage;
         Long lastPage = (orderQuantity % quantityOnPage != 0 ? pagesQuantity + 1 : pagesQuantity);
         model.addAttribute("lastPage", lastPage);
-        model.addAttribute("orders", orderDao.findAll((int) ((page - 1) * quantityOnPage), quantityOnPage.intValue()));
+        model.addAttribute("orders", orderService.findAllOrders((int) ((page - 1) * quantityOnPage),
+                quantityOnPage.intValue()));
         return "adminOrders";
     }
 }
