+++ web/src/main/webapp/WEB-INF/securityContext.xml
@@ -8,7 +8,6 @@
 
     <http auto-config="true">
         <intercept-url pattern="/admin/*" access="ROLE_ADMIN" />
-
         <form-login
                 login-page="/login"
                 default-target-url="/productList"
@@ -16,15 +15,17 @@
                 username-parameter="username_login"
                 password-parameter="password_login"
         />
-
         <logout logout-success-url="/login"/>
     </http>
 
     <authentication-manager>
         <authentication-provider>
-            <user-service>
-                <user name="admin" password="admin" authorities="ROLE_USER, ROLE_ADMIN" />
-            </user-service>
+            <jdbc-user-service
+                    data-source-ref="dataSource"
+                    users-by-username-query="select username, password, enabled from users where username = ?"
+                    authorities-by-username-query="select u.username, au.authority from users u, authorities au
+       	                                                where u.id = au.userId and u.username = ?"
+            />
         </authentication-provider>
     </authentication-manager>
